+++ modele/make_topoa.cpp
@@ -220,9 +220,9 @@ int main(int argc, char **argv)
         get_or_put_att(info, topoa_nc.rw, "segments", sval);
 
         ncio_vector(topoa_nc, lonc, false, "lon", "double",
-            get_or_add_dims(topoa_nc, {"im"}, {lonc.size()}));
+            get_or_add_dims(topoa_nc, {"im"}, {(long)lonc.size()}));
         ncio_vector(topoa_nc, latc, false, "lat", "double",
-            get_or_add_dims(topoa_nc, {"jm"}, {latc.size()}));
+            get_or_add_dims(topoa_nc, {"jm"}, {(long)latc.size()}));
 
 
         // Write topoA arrays
